requirements.txt
altair==5.3.0
ipeadatapy==0.1.9
matplotlib==3.9.0
numpy==1.26.0
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
scikit-learn==1.5.1
streamlit==1.34.0
